implementing grade functionaliteies

+++ Server/Scholario.Application/Dtos/Grade/AddOrUpdateGradeToStudentDto.cs
@@ -1,16 +1,11 @@
 ﻿using Scholario.Domain.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Scholario.Application.Dtos.Grade
 {
     public class AddOrUpdateGradeToStudentDto
     {
         public int? Id { get; set; }
-        public int GradeValue { get; set; }
+        public float GradeValue { get; set; }
         public GradeWeight GradeWeight { get; set; }
         public int SubjectId { get; set; }
         public int StudentId { get; set; }
